Address warnings from clang build bot
http://codereview.appspot.com/6506111/



git-svn-id: http://skia.googlecode.com/svn/trunk@5518 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/core/SkRTree.cpp
@@ -399,7 +399,7 @@ void SkRTree::validate() {
     if (this->isEmpty()) {
         return;
     }
-    SkASSERT(fCount == this->validateSubtree(fRoot.fChild.subtree, fRoot.fBounds, true));
+    SkASSERT(fCount == (size_t)this->validateSubtree(fRoot.fChild.subtree, fRoot.fBounds, true));
 #endif
 }
